Added base classes to remove duplicate code
