+++ lib/protocol/base.js
@@ -8,10 +8,14 @@
 const datapoints = require(__dirname + '/../datapoints');
 
 class BaseClient {
-    constructor(adapter, objectHelper, eventEmitter) {
+    constructor(type, adapter, objectHelper, eventEmitter) {
+        this.type = type;
+
         this.adapter = adapter;
         this.objectHelper = objectHelper;
         this.eventEmitter = eventEmitter;
+
+        this.timerid = null;
 
         this.active = true;
         this.states = {};
@@ -19,8 +23,8 @@ class BaseClient {
         this.http = {};
 
         this.devicename;
-        this.ip;
-        this.auth;
+        this.ip;            // e.g. 192.168.1.2
+        this.auth;          // Authorization HTTP Header
 
         this.id;
         this.devicetype;
@@ -91,9 +95,137 @@ class BaseClient {
         return this.devicegen;
     }
 
+    /**
+    * Missing data will be pulled by http
+    */
+    async httpIoBrokerState() {
+        // Test - to delete after a while
+        if (!this.httpIoBrokerStateTime || Date.now() >= (this.httpIoBrokerStateTime + (1000 * 60 * 10))) {
+            this.httpIoBrokerStateTime = Date.now();
+        }
+
+        if (!this.adapter.isOnline(this.getDeviceId())) {
+            this.timerid = setTimeout(async () => await this.httpIoBrokerState(), 100);
+            return;
+        }
+
+        let polltime = this.getPolltime();
+        for (let i in this.http) {
+            let params;
+            let states = this.http[i];
+            try {
+                if (this.auth) {
+                    params = {
+                        url: 'http://' + this.getIP() + i,
+                        timeout: this.httptimeout,
+                        headers: {
+                            'Authorization': this.auth
+                        }
+                    };
+                } else {
+                    params = {
+                        url: 'http://' + this.getIP() + i,
+                        timeout: this.httptimeout
+                    };
+                }
+
+                this.adapter.log.debug('http request' + JSON.stringify(params) + ' for ' + this.getName());
+
+                let body = await requestAsync(params);
+                for (let j in states) {
+                    let state = this.device[states[j]];
+                    if (state && state.state) {
+                        let deviceid = this.getDeviceName();
+                        let stateid = deviceid + '.' + state.state;
+                        let value = body;
+                        try {
+                            if (state[this.type] && state[this.type].http_publish_funct) {
+                                value = isAsync(state[this.type].http_publish_funct) ? await state[this.type].http_publish_funct(value, this) : state[this.type].http_publish_funct(value, this);
+                            }
+
+                            if (state.common.type === 'boolean' && value === 'false') value = false;
+                            if (state.common.type === 'boolean' && value === 'true') value = true;
+                            if (state.common.type === 'number' && value !== undefined) value = Number(value);
+
+                            if (value !== undefined && (!Object.prototype.hasOwnProperty.call(this.states, stateid) || this.states[stateid] !== value || this.adapter.config.updateUnchangedObjects)) {
+                                this.adapter.log.debug('Set http state ' + stateid + ', Value: ' + JSON.stringify(value) + ' for ' + this.getName());
+                                this.states[stateid] = value;
+                                this.objectHelper.setOrUpdateObject(stateid, {
+                                    type: 'state',
+                                    common: state.common
+                                }, ['name'], value);
+                            }
+                        } catch (error) {
+                            if (error.name && error.name.startsWith('TypeError')) {
+                                this.adapter.log.debug('Could not find property for state ' + stateid + ' and request' + JSON.stringify(params) + ' for ' + this.getName() + ' (' + error + ')');
+                            } else {
+                                if (polltime > 0 && polltime < 60) polltime = 60;
+                                if (body && body === '401 Unauthorized') {
+                                    this.adapter.log.error('Wrong http username or http password! Please enter the user credential from restricted login for ' + this.getName());
+                                    break;
+                                } else {
+                                    this.adapter.log.error('Error in function httpIoBrokerState for state ' + stateid + ' and request' + JSON.stringify(params) + ' for ' + this.getName() + ' (' + error + ')');
+                                }
+                            }
+                        }
+                    }
+                }
+                this.objectHelper.processObjectQueue(() => { });
+            } catch (error) {
+                //
+            }
+        }
+
+        if (this.http && Object.keys(this.http).length > 0 && polltime > 0) {
+            this.timerid = setTimeout(async () => await this.httpIoBrokerState(), polltime * 1000);
+        }
+    }
+
+    /*
+    async firmwareUpdatePolling() {
+        if (this.adapter.config.autoupdate) {
+            await this.firmwareUpdate(true);
+            this.autoupdateid = setTimeout(async () => await this.firmwareUpdatePolling(), 60 * 1000);
+        }
+    }
+    */
+
+    async firmwareUpdate(update) {
+        if (!update) return;
+
+        this.adapter.log.debug('Calling function firmwareUpdate');
+        let params;
+
+        try {
+            if (this.auth) {
+                params = {
+                    url: 'http://' + this.getIP() + '/ota?update=true',
+                    timeout: this.httptimeout,
+                    headers: {
+                        'Authorization': this.auth
+                    }
+                };
+            } else {
+                params = {
+                    url: 'http://' + this.getIP() + '/ota?update=true',
+                    timeout: this.httptimeout
+                };
+            }
+
+            this.adapter.log.debug('Call url ' + JSON.stringify(params) + ' for ' + this.getName());
+            let body = await requestAsync(params);
+
+            this.adapter.log.debug('Executing Firmwareupdate for ' + this.getName());
+        } catch (error) {
+            this.adapter.log.error('Error in function firmwareUpdate and request' + JSON.stringify(params) + ' for ' + this.getName() + ' (' + error + ')');
+        }
+    }
+
     destroy() {
         if (this.active) {
             this.adapter.log.debug('Destroy ' + this.getName());
+
+            clearTimeout(this.timerid);
 
             this.active = false;
             this.states = {};
@@ -109,7 +241,8 @@ class BaseClient {
             this.deviceid = undefined;
             this.serialid = undefined;
             this.deviceexist = undefined;
-            this.devicegen = undefined;        }
+            this.devicegen = undefined;
+        }
     }
 }
