+++ lib/protocol/coap.js
@@ -10,7 +10,6 @@ const BaseServer = require(__dirname + '/base').BaseServer;
 const datapoints = require(__dirname + '/../datapoints');
 
 const Shelly = require('shelly-iot');
-const tcpp = require('tcp-ping');
 const request = require('request');
 
 function sleep(ms) {
@@ -20,20 +19,6 @@ function sleep(ms) {
 function isAsync(funct) {
   if (funct && funct.constructor) return funct.constructor.name == 'AsyncFunction';
   return undefined;
-}
-
-/**
- * Ping host with tcp
- * @param {*} host - hostname like www.google.de or 192.168.20.1
- * @param {*} port - 80 or 443
- */
-function pingAsync(host, port) {
-  if (!port) port = 80;
-  return new Promise((resolve, reject) => {
-    tcpp.probe(host, port, (error, isAlive) => {
-      resolve(isAlive);
-    });
-  });
 }
 
 function requestAsync(url) {
@@ -111,7 +96,7 @@ function descrToSensor(description) {
 class CoAPClient extends BaseClient {
   constructor(adapter, objectHelper, eventEmitter, shelly, devicename, ip, payload, description) {
 
-    super(adapter, objectHelper, eventEmitter);
+    super('coap', adapter, objectHelper, eventEmitter);
 
     this.getOldDeivceInfo(devicename);
 
@@ -213,9 +198,6 @@ class CoAPClient extends BaseClient {
    */
   destroy() {
     super.destroy();
-
-    clearTimeout(this.timerid);
-    clearTimeout(this.autoupdateid);
 
     this.description = undefined;
 
@@ -436,122 +418,6 @@ class CoAPClient extends BaseClient {
       }
     }
     this.objectHelper.processObjectQueue(() => { });
-  }
-
-  /**
-   * Missing data in CoAP will be pulled by http
-   */
-  async httpIoBrokerState() {
-    // Test - to delete after a while
-    if (!this.httpIoBrokerStateTime || Date.now() >= (this.httpIoBrokerStateTime + (1000 * 60 * 10))) {
-      this.httpIoBrokerStateTime = Date.now();
-      // this.adapter.log.info('httpIoBrokerState() still running on ' + this.getName() + ' for ' + JSON.stringify(this.http));
-    }
-    let alive = await pingAsync(this.getIP());
-    if (alive === false) {
-      this.timerid = setTimeout(async () => await this.httpIoBrokerState(), 100);
-      return;
-    }
-    let polltime = this.getPolltime();
-    for (let i in this.http) {
-      let params;
-      let states = this.http[i];
-      try {
-        if (this.auth) {
-          params = {
-            url: 'http://' + this.getIP() + i,
-            timeout: this.httptimeout,
-            headers: {
-              'Authorization': this.auth
-            }
-          };
-        } else {
-          params = {
-            url: 'http://' + this.getIP() + i,
-            timeout: this.httptimeout
-          };
-        }
-        this.adapter.log.debug('http request' + JSON.stringify(params) + ' for ' + this.getName());
-        let body = await requestAsync(params);
-        for (let j in states) {
-          let state = this.device[states[j]];
-          if (state && state.state) {
-            let deviceid = this.getDeviceName();
-            let stateid = deviceid + '.' + state.state;
-            let value = body;
-            try {
-              if (state.coap && state.coap.http_publish_funct)
-                value = isAsync(state.coap.http_publish_funct) ? await state.coap.http_publish_funct(value, this) : state.coap.http_publish_funct(value, this);
-              if (state.common.type === 'boolean' && value === 'false') value = false;
-              if (state.common.type === 'boolean' && value === 'true') value = true;
-              if (state.common.type === 'number' && value !== undefined) value = Number(value);
-              if (value !== undefined && (!Object.prototype.hasOwnProperty.call(this.states, stateid) || this.states[stateid] !== value || this.adapter.config.updateUnchangedObjects)) {
-                this.adapter.log.debug('Set http state ' + stateid + ', Value: ' + JSON.stringify(value) + ' for ' + this.getName());
-                this.states[stateid] = value;
-                this.objectHelper.setOrUpdateObject(stateid, {
-                  type: 'state',
-                  common: state.common
-                }, ['name'], value);
-              }
-            } catch (error) {
-              if (error.name && error.name.startsWith('TypeError')) {
-                this.adapter.log.debug('Could not find property for state ' + stateid + ' and request' + JSON.stringify(params) + ' for ' + this.getName() + ' (' + error + ')');
-              } else {
-                if (polltime > 0 && polltime < 60) polltime = 60;
-                if (body && body === '401 Unauthorized') {
-                  this.adapter.log.error('Wrong http username or http password! Please enter the user credential from restricted login for ' + this.getName());
-                  break;
-                } else {
-                  this.adapter.log.error('Error in function httpIoBrokerState for state ' + stateid + ' and request' + JSON.stringify(params) + ' for ' + this.getName() + ' (' + error + ')');
-                }
-              }
-            }
-          }
-        }
-        this.objectHelper.processObjectQueue(() => { });
-      } catch (error) {
-        //
-      }
-    }
-    if (this.http && Object.keys(this.http).length > 0 && polltime > 0) {
-      this.timerid = setTimeout(async () => await this.httpIoBrokerState(), polltime * 1000);
-    }
-  }
-
-  /*
-  async firmwareUpdatePolling() {
-    if (this.adapter.config.autoupdate) {
-      await this.firmwareUpdate(true);
-      this.autoupdateid = setTimeout(async () => await this.firmwareUpdatePolling(), 60 * 1000);
-    }
-  }
-  */
-
-  async firmwareUpdate(update) {
-    if (!update) return;
-    this.adapter.log.debug('Calling function firmwareUpdate');
-    let params;
-    try {
-      if (this.auth) {
-        params = {
-          url: 'http://' + this.getIP() + '/ota?update=true',
-          timeout: this.httptimeout,
-          headers: {
-            'Authorization': this.auth
-          }
-        };
-      } else {
-        params = {
-          url: 'http://' + this.getIP() + '/ota?update=true',
-          timeout: this.httptimeout
-        };
-      }
-      this.adapter.log.debug('Call url ' + JSON.stringify(params) + ' for ' + this.getName());
-      let body = await requestAsync(params);
-      // this.adapter.log.info('Executing Firmwareupdate for ' + this.getName());
-    } catch (error) {
-      this.adapter.log.error('Error in function firmwareUpdate and request' + JSON.stringify(params) + ' for ' + this.getName() + ' (' + error + ')');
-    }
   }
 
   start(payload, description) {
